refactor: Move IPC message classes into separate files
* Moving the classes out of the NamedPipeMessages static class and removing that class
