+++ src/CoreHook.IPC/Messages/ICustomMessage.cs
@@ -1,0 +1,10 @@
+﻿
+namespace CoreHook.IPC.Messages
+{
+    public abstract class CustomMessage
+    {
+        protected const char MessageSeparator = '|';
+
+        public abstract string ToMessage();
+    }
+}
