+++ src/CoreHook.IPC/Messages/IMessage.cs
@@ -1,0 +1,12 @@
+﻿
+namespace CoreHook.IPC.Messages
+{
+    public interface IMessage
+    {
+        string Header { get; }
+
+        string Body { get; }
+
+        string ToMessage();
+    }
+}
